+++ gixy/plugins/low_keepalive_requests.py
@@ -12,7 +12,10 @@ class low_keepalive_requests(Plugin):
 
     summary = "The keepalive_requests directive should be at least 1000."
     severity = gixy.severity.LOW
-    description = "The keepalive_requests directive should be at least 1000. Any value lower than this may result in client disconnections."
+    description = (
+        "The keepalive_requests directive should be at least 1000. Any value "
+        "lower than this may result in client disconnections."
+    )
     directives = ["keepalive_requests"]
 
     def audit(self, directive):